% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{calc_sub_graph}
\alias{calc_sub_graph}
\title{The nodes within the graph (or the subgraph) before cutting the max_edge. This function
aims to mark all the graph nodes before cutting the max_edge.}
\usage{
calc_sub_graph(cur)
}
\arguments{
\item{cur}{The current node}
}
\value{
The function returns the nodes contained within the graph (subgraph) before cutting the max_edge.
calc_subgraph(cur)
}
\description{
The nodes within the graph (or the subgraph) before cutting the max_edge. This function
aims to mark all the graph nodes before cutting the max_edge.
}
