% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{dfs}
\alias{dfs}
\title{The contained nodes in the sugraph after cutting the max_edge.
This function visits and mark the nodes of the subgraph after cutting the max_edge.}
\usage{
dfs(cur)
}
\arguments{
\item{cur}{The current node}
}
\value{
It uses the Depth First Search algorithm to return the visited  nodes after cutting the max_edge in order to check whether the subgraph remains connected after the cut or not.
dfs(currentnode)
}
\description{
The contained nodes in the sugraph after cutting the max_edge.
This function visits and mark the nodes of the subgraph after cutting the max_edge.
}
