% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{gbc}
\alias{gbc}
\title{Gbc function is a graph based clustering method.
A function to detect the appropriate number of clusters , and returns the appropriate clustering.}
\usage{
gbc(x)
}
\arguments{
\item{x}{Any given dataset, any (.txt, .csv,..ect.),
such that the separator between its columns and lines should be a blank space " ",
and the number of its observations (lines) and the number of its features should be
a moderate number (not too large), and the type of its elements (attributes) should
be of numerical type (continous or multivalued values) for calculating the distances.
and the observations with missing values should be removed from the dataset (and note
that you can use the Milligan Cooper method to standardize the dataset's values).}
}
\value{
This algorithm detects the appropriate number of clusters k, so the appropriate clustering.
}
\description{
Gbc function is a graph based clustering method.
A function to detect the appropriate number of clusters , and returns the appropriate clustering.
}
