% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{rng}
\alias{rng}
\title{Draw the Neighborhood Graph-NG of the dataset}
\usage{
rng(dataset)
}
\arguments{
\item{dataset}{Any given dataset, any (.txt, .csv,..ect.),
such that the separator between its columns and lines should be a blank space " ",
and the number of its observations (lines) and the number of its features should be
a moderate number (not too large), and the type of its elements (attributes) should
be of numerical type (continous or multivalued values) for calculating the distances.
and the observations with missing values should be removed from the dataset (and note
that you can use the Milligan Cooper method to standardize the dataset's values).}
}
\value{
The function plots the NG graph (RNG) and returns the sorted list by size(taille/poids) in decreasing order of the created edges of the RNG graph.
rng(dataset)
}
\description{
Rng a function to construct the NG graph (Neighborhood Graph) using
the relative neighborhood graph structure, in our case(RNG).
}
