\name{ruspini}
\alias{ruspini}
\docType{data}
\title{coordinates x and y of 75 points in the abscissa and ordinate axes}
\description{
The Ruspini data set, consisting of 75 points in four groups that is popular for illustrating clustering techniques.
}
\usage{data("ruspini")}
\format{
  A data frame with 75 observations on the following 2 variables.
  \describe{
    \item{\code{X4}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
  }
}
\source{
https://stat.ethz.ch/R-manual/R-devel/library/cluster/html/ruspini.html
}
\examples{
data(ruspini)
## maybe str(ruspini) ; plot(ruspini) ...
}
\keyword{datasets}
