% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{set_cluster_number}
\alias{set_cluster_number}
\title{The number of the cluster for each node.
This function associates each node to its cluster (where it belongs).}
\usage{
set_cluster_number(cur)
}
\arguments{
\item{cur}{The current node}
}
\value{
It returns the number of the cluster of each node in the dataset.
}
\description{
The number of the cluster for each node.
This function associates each node to its cluster (where it belongs).
}
